Create Client DAO and DTO
