+++ app/src/main/java/com/example/appvolk/Activity/CadastroAtividadesActivity.java
@@ -1,24 +1,39 @@
-package com.example.appvolk;
+package com.example.appvolk.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.app.Activity;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
 
-public class CadastroSocioActivity extends AppCompatActivity {
+import com.example.appvolk.R;
+
+public class CadastroAtividadesActivity extends AppCompatActivity {
+
+    TextView textVoltar3;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_cadastro_socio);
+        setContentView(R.layout.activity_cadastro_atividades);
+
+        textVoltar3 = findViewById(R.id.textVoltar3);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
+        textVoltar3();
 
+    }
+
+    private void textVoltar3(){
+        textVoltar3.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(),MenuActivity.class);
+                startActivity(intent);
+            }
+        });
     }
 
     public boolean onOptionsItemSelected(MenuItem menuItem){
@@ -30,7 +45,6 @@ public class CadastroSocioActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(menuItem);
     }
+
+
 }
-
-
-
