+++ app/src/main/java/com/example/appvolk/Activity/CadastroSocioActivity.java
@@ -1,37 +1,26 @@
-package com.example.appvolk;
+package com.example.appvolk.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.app.Activity;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
 
-public class SobreActivity extends AppCompatActivity {
+import com.example.appvolk.R;
 
-    TextView textVoltar9;
+public class CadastroSocioActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sobre);
-
-        textVoltar9 = findViewById(R.id.textVoltar9);
+        setContentView(R.layout.activity_cadastro_socio);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
-        textVoltar9();
 
-    }
-
-    private void textVoltar9(){
-        textVoltar9.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(),MenuActivity.class);
-                startActivity(intent);
-            }
-        });
     }
 
     public boolean onOptionsItemSelected(MenuItem menuItem){
@@ -43,5 +32,7 @@ public class SobreActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(menuItem);
     }
-
 }
+
+
+
