+++ app/src/main/java/com/example/appvolk/Activity/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.appvolk;
+package com.example.appvolk.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+
+import com.example.appvolk.R;
 
 public class LoginActivity extends AppCompatActivity {
