+++ app/src/main/java/com/example/appvolk/Database/ClienteDTO.java
@@ -1,0 +1,49 @@
+package com.example.appvolk.Database;
+
+public class ClienteDTO {
+    private int id;
+    private String cpf, nome, email, telefone;
+
+    public ClienteDTO() {
+
+    }
+
+    public ClienteDTO(String cpf, String nome, String email, String telefone) {
+        this.cpf = cpf;
+        this.nome = nome;
+        this.email = email;
+        this.telefone = telefone;
+    }
+
+    public int getCPF() {
+        return id;
+    }
+
+    public void setCPF(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public String getNome() {
+        return this.nome;
+    }
+
+    public void setNome(String name) {
+        this.nome = nome;
+    }
+
+    public String getEmail() {
+        return this.email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getTelefone() {
+        return this.telefone;
+    }
+
+    public void setTelefone(String telefone) {
+        this.telefone = telefone;
+    }
+}
