+++ app/src/main/res/layout/activity_cadastro_cliente.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@mipmap/ic_fundo_foreground"
-    tools:context=".CadastroClienteActivity">
+    tools:context=".Activity.CadastroClienteActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
