+++ app/src/main/res/layout/activity_consulta_atividades.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@mipmap/ic_fundo_foreground"
-    tools:context=".ConsultaAtividadesActivity">
+    tools:context=".Activity.ConsultaAtividadesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
