+++ app/src/main/res/layout/activity_consulta_clientes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@mipmap/ic_fundo_foreground"
-    tools:context=".ConsultaClientesActivity">
+    tools:context=".Activity.ConsultaClientesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
