+++ app/src/main/res/layout/activity_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@mipmap/ic_fundo_foreground"
-    tools:context=".MenuActivity">
+    tools:context=".Activity.MenuActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
