+++ app/src/main/res/layout/activity_sobre.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@mipmap/ic_fundo_foreground"
-    tools:context=".SobreActivity">
+    tools:context=".Activity.SobreActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
